customer api working

+++ springboottraining/src/main/java/com/siemens/springboottraining/service/CustomerService.java
@@ -40,8 +40,11 @@ public class CustomerService implements ICustomerService {
     @Override
     public boolean deleteCustomer(long customerId) {
         Customer customer = this.getCustomerById(customerId);
-        this.customerRepository.delete(customer);
-        return this.getCustomerById(customerId) == null;
+        if (customer != null) {
+            this.customerRepository.delete(customer);
+            return true;
+        }
+        return false;
     }
 
     @Override
